% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bowhead_LT}
\alias{bowhead_LT}
\title{Line transect survey data from bowhead whale survey.}
\format{
A data frame of line transect data suitable for analysis with package \code{Distance}.
}
\source{
Rekdal, S.L., Hansen, R.G., Borchers, D., Bachmann, L., Laidre, K.L., Wiig, O., Nielsen, N.H., 
Fossette, S., Tervo, O. and Heide-Jorgensen, M.P., 2015. Trends in bowhead whales in West Greenland: 
Aerial surveys vs. genetic capture-recapture analyses. Marine Mammal Science, 31(1), pp.133-154.
}
\usage{
bowhead_LT
}
\description{
Line transect survey data from the bowhead whale survey reported in Rekdal et al. (2015).
}
\examples{
 data(bowhead_LT)
}
