% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bowhead_PS}
\alias{bowhead_PS}
\title{Plot sampling survey data from Bowhead whale survey.}
\format{
A data frame with one row per plot and columns \code{stratum}, \code{transect} (plot), \code{A} 
(stratum area), \code{a} (plot area), and \code{n} (count of whale groups in plot).
}
\source{
Rekdal, S.L., Hansen, R.G., Borchers, D., Bachmann, L., Laidre, K.L., Wiig, O., Nielsen, N.H., 
Fossette, S., Tervo, O. and Heide-Jorgensen, M.P., 2015. Trends in bowhead whales in West Greenland: 
Aerial surveys vs. genetic capture-recapture analyses. Marine Mammal Science, 31(1), pp.133-154.
}
\usage{
bowhead_PS
}
\description{
Data from the bowhead whale survey reported in Rekdal et al. (2015), but formatted as 
a plot sampling survey.
}
\examples{
 data(bowhead_PS)
}
