% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crfunctions.R
\name{delta.var}
\alias{delta.var}
\title{Delta method variance}
\usage{
delta.var(x, var.x, link)
}
\arguments{
\item{x}{Random variable that inverse transformation is applied to}

\item{var.x}{Variance of \code{x}.}
}
\value{
The Delta Method (first order Taylor series) approximation.
}
\description{
Calculated Delta Method variance for inverse log and inverse logistic link functions.
}
