% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{edwards.eberhardt}
\alias{edwards.eberhardt}
\title{Rabbit capture-recapture data}
\format{
A capture history matrix of 76 rabbits (rows) on 18 occasions (columns), with 0 representing
not captured and 1 representing captured.
}
\source{
Edwards, W.R. and L.L. Eberhardt 1967. Estimating cottontail abundance from live trapping data. J. Wildl. Manage. 31:87-96.
}
\usage{
edwards.eberhardt
}
\description{
A capture-recapture data set on rabbits derived from Edwards and Eberhardt (1967) that accompanies 
programme MARK as an example analysis using the closed population models.
}
\examples{
 data(edwards.eberhardt)
}
