% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crfunctions.R
\name{loglik.Mh2}
\alias{loglik.Mh2}
\title{Negative log-likelihood for model Mh2}
\usage{
loglik.Mh2(pars, chmat)
}
\arguments{
\item{pars}{Parameter vector comprising (qlogis(p1), qlogis(p2), qlogis(phi)), where
p1 and p2 are the two latent capture probabilities, and phi is the proportion of the population
with capture probability p1.}

\item{chmat}{Binary capture history matrix, with one row per individual detected and 
one column per capture occasion.}
}
\value{
The negative log-likelihood.
}
\description{
Calculates the negative log-likelihood for a 2-part finite mixture model.
}
