% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crfunctions.R
\name{loglik.Mtb}
\alias{loglik.Mtb}
\title{Negative log-likelihood for models M0, Mt, Mb, Mtb}
\usage{
loglik.Mtb(pars, chmat, prevcap = NULL, timecov = NULL)
}
\arguments{
\item{pars}{Parameter vector with the correct number of parameters in the correct order 
(see Details).}

\item{chmat}{Binary capture history matrix, with one row per individual detected and 
one column per capture occasion.}

\item{prevcap}{Binary matrix of same size as \code{chmat}, with 0 indicating that the
individual has not been captured by that occasion, and 1 indicating that it has. Used
by the function to decide what the model is, so if it is not null, the model is taken
to be Mb (if \code{timecov} is NULL) or Mtb (if \code{timecov} is not NULL).}

\item{timecov}{Matrix of same size as \code{chmat} indicating the time associated with the
capture occasion. At present it is treated as a factor.}
}
\value{
The negative log-likelihood for the appropriate model.
}
\description{
Calculates the negative log-likelihood for models M0, Mt, Mb, Mtb.
}
\details{
The function decides which model to use on the basis of which of \code{prevcap} and 
\code{timecov} are null. It assumes that \code{pars} has the appropriate number of parameters
for the model, and in the appropriate order, as follows:
\itemize{
 \item{'M0'}{ \code{prevcap} and \code{timecov} both NULL: \code{pars} must be 
 \code{(log(N),qlogis(p))}}.
 \item{'Mt'}{ \code{prevcap} NULL, \code{timecov} not: \code{pars} must be 
 \code{(log(N),qlogis(p1)...qlogis(pK))}, where there are \code{K} occasions.}
 \item{'Mb'}{ \code{prevcap} not NUL, \code{timecov} is NULL: \code{pars} must be 
 \code{(log(N),qlogis(p),log(b))}, where \code{log(b)} is the change in log-odds due to 
 previous capture.}
 \item{'Mtb'}{ Neither \code{prevcap} nor \code{timecov} are NULL: \code{pars} must be 
 \code{(log(N),qlogis(p1)...qlogis(pK),log(b))}, where there are \code{K} occasions and 
 \code{log(b)} is the change in log-odds due to previous capture.}
}
}
