% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lognorm.ci}
\alias{lognorm.ci}
\title{Lognormal confidence interval}
\usage{
lognorm.ci(x, se, alpha = 0.05)
}
\arguments{
\item{x}{The estimate}

\item{se}{The standard error of the estimate}

\item{alpha}{The significance level (defaults to \code{alpha}=0.05 for 95\% level)}
}
\value{
The function returns a list with the following elements:
}
\description{
Returns confidence interval assuming estimator has a lognormal distribution
}
