% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crfunctions.R
\name{make.pests}
\alias{make.pests}
\title{Make matrix of probability estimates}
\usage{
make.pests(parnames, ests, vcv, b = FALSE, alpha = 0.05)
}
\arguments{
\item{parnames}{Names of the parameters in \code{ests}.}

\item{ests}{Vector of parameter estimates on the link function scale.}

\item{vcv}{The variance-covariance matrix of \code{ests}.}

\item{b}{Indicator variable saying whether or not the last parameter in \code{ests} is
the behavioural effect parameter, in which case it is interpreted as change in log-odds.}

\item{alpha}{Confidence interval level.}
}
\value{
The function returns a matrix with one row per parameter and columns for the point estimate
on the natural scale and associated confidence intervals.
}
\description{
Make matrix of probability estimates on the natural scale, from estimates on the linke
function scale, and their variance-covariance matrix.
}
