% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crfunctions.R
\name{make.prevcap}
\alias{make.prevcap}
\title{Make binary covariate for previous capture}
\usage{
make.prevcap(chmat)
}
\arguments{
\item{chmat}{Capture history matrix.}
}
\value{
The function returns a matrix with one row per individual, one column per occasion, with 
a 0 indicating the animal had not been captured before an a 1 indicating that it had.
}
\description{
Makes an indicator variable for whether or not an animal has been captured before
each occasion.
}
