% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt5751a-package.r
\docType{package}
\name{mt5751a-package}
\alias{mt5751a-package}
\alias{mt5751a}
\title{Functions for first part of MT5751}
\description{
Plot sampling and capture-recapture functions for use in MT5751.
}
\details{
\tabular{ll}{ Package: \tab mt5751a \cr Type: \tab Package \cr 
  Version: \tab 1.0 \cr Date: \tab 2021-01-30 \cr License: \tab GNU General 
  Public License Version 2 or later \cr }
}
\author{
David Borchers \email{dlb@st-andrews.ac.uk}
  
  Maintainer: David Borchers \email{dlb@st-andrews.ac.uk}
}
