% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsampling.R
\name{norm.ci}
\alias{norm.ci}
\title{Normal confidence interval}
\usage{
norm.ci(x, se, alpha)
}
\arguments{
\item{x}{The estimate}

\item{se}{The standard error of the estimate}

\item{alpha}{The significance level (defaults to \code{alpha}=0.05 for 95\% level)}
}
\value{
stuff
}
\description{
Returns confidence interval assuming estimator has a normal distribution
}
