% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{plotsample2}
\alias{plotsample2}
\title{Simulated plot sample data.}
\format{
A data frame with one row per plot and columns \code{stratum}, \code{transect} (plot), \code{A} 
(stratum area), \code{a} (plot area), and \code{n} (count of whale groups in plot).
}
\source{
Simulated
}
\usage{
plotsample2
}
\description{
Simulated plot sample data from 20 plots in a 100m by 100m survey region.
}
\examples{
 data(plotsample2)
}
