% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tigerch}
\alias{tigerch}
\title{Simulated capture-recapture tiger survey dataset.}
\format{
A capture history matrix of 56 animals (rows) on 5 occasions (columns), with 0 representing
not captured and 1 representing captured.
}
\source{
Simulated
}
\usage{
tigerch
}
\description{
Simulated capture history from a simulated population of 81 tigers in the Sundarbans
region of India. Simulation is based on a real camera trap survey of the tigers. (Object 
\link{tigersecrch} contains the same data, but with spatial capture information.)
}
\examples{
 data(tigerch)
}
