% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tigermask}
\alias{tigermask}
\title{The \code{secr} \code{mask} object for simualted tiger survey.}
\format{
An \code{secr} \code{mask} object defining the survey region.
}
\source{
Simulated
}
\usage{
tigermask
}
\description{
The \code{secr} \code{mask} object for the \code{tigersecrech} object.
}
\examples{
 data(tigermask)
}
