% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tigerpop}
\alias{tigerpop}
\title{A simulated tiger population.}
\format{
An \code{secr} \code{mask} object defining the survey region.
}
\source{
Simulated
}
\usage{
tigerpop
}
\description{
The \code{secr} \code{popn} object defining the population that was 
surveyed to get the capture history \code{tigersecrch}.
}
\examples{
 data(tigerpop)
}
