% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tigersecrch}
\alias{tigersecrch}
\title{Simulated spatial capture-recapture tiger survey dataset.}
\format{
An \code{secr} \code{capthist} object containing 56 animals on 5 occasions.
}
\source{
Simulated
}
\usage{
tigersecrch
}
\description{
Simulated spatial capture history from a simulated population of 81 tigers in the Sundarbans
region of India. Simulation is based on a real camera trap survey of the tigers. (Object 
\link{tigerch} contains the same capture history data, but with spatial capture information removed.)
}
\examples{
 data(tigersecrch)
}
\seealso{
\code{\link{tigermask}}
}
